Shard P2MR wallet unit tests
